+++ apps/api/src/core/analytics/analytics.controller.ts
@@ -2,20 +2,17 @@ import { Controller, Get, Param, Query, UseGuards } from "@nestjs/common";
 import { ApiBearerAuth, ApiQuery, ApiTags } from "@nestjs/swagger";
 
 import { JwtAuthGuard } from "../../auth/guards/jwt-auth.guard";
-import { LogDto } from "../log/dto/log.dto";
-import { LogService } from "../log/log.service";
 import { DateSelection } from "./analytics.enum";
 import { AnalyticsService } from "./analytics.service";
 import { DashboardAnalyticsDto } from "./dto/dashboard-analytics.dto";
+import { CostHistoryDto } from "./dto/cost-history.dto";
+
 
 @ApiTags("Analytics")
 @ApiBearerAuth()
 @Controller("env/:envId/analytics")
 export class AnalyticsController {
-  constructor(
-    private analyticsService: AnalyticsService,
-    private logService: LogService,
-  ) {}
+  constructor(private analyticsService: AnalyticsService) {}
 
   @ApiQuery({ name: "dateSelection", enum: DateSelection })
   @UseGuards(JwtAuthGuard)
@@ -24,17 +21,23 @@ export class AnalyticsController {
     @Param("envId") envId: string,
     @Query("dateSelection") dateSelection: DateSelection,
   ): Promise<DashboardAnalyticsDto> {
-    const dashboardAnalyticsPromise = this.analyticsService.getDashboardAnalytics({
+    return await this.analyticsService.getDashboardAnalytics({
       envId,
       dateSelection,
     });
-    const recentLogsPromise = this.logService.findAllTopLevel(envId, { take: 50 });
+  }
 
-    const [analytics, logs] = await Promise.all([dashboardAnalyticsPromise, recentLogsPromise]);
-    const logsDto = logs.map(LogDto.fromLog);
-    return {
-      ...analytics,
-      logs: logsDto,
-    };
+  @ApiQuery({ name: "dateSelection", enum: DateSelection })
+  @UseGuards(JwtAuthGuard)
+  @Get("cost-history")
+  async costHistory(
+    @Param("envId") envId: string,
+    @Query("dateSelection") dateSelection: DateSelection,
+  ): Promise<CostHistoryDto> {
+    const costHistory = await this.analyticsService.getCostHistory({
+      envId,
+      dateSelection,
+    });
+    return { costHistory };
   }
 }