+++ apps/api/src/core/analytics/analytics.module.ts
@@ -1,12 +1,12 @@
 import { Module } from "@nestjs/common";
 
 import { DatabaseModule } from "../../database";
-import { LogModule } from "../log/log.module";
 import { AnalyticsController } from "./analytics.controller";
 import { AnalyticsService } from "./analytics.service";
 
+
 @Module({
-  imports: [DatabaseModule, LogModule],
+  imports: [DatabaseModule],
   controllers: [AnalyticsController],
   providers: [AnalyticsService],
   exports: [AnalyticsService],