+++ apps/api/src/core/analytics/analytics.types.ts
@@ -1,5 +1,15 @@
 import { DateSelection } from "./analytics.enum";
 
+export type GetCostHistoryParams = {
+  envId: string;
+  dateSelection: DateSelection;
+};
+
+export type CostHistory = {
+  intervalStart: string;
+  totalCost: number;
+};
+
 export type GetDashboardAnalyticsParams = {
   envId: string;
   dateSelection: DateSelection;