+++ apps/api/src/core/analytics/dto/cost-history.dto.ts
@@ -1,0 +1,15 @@
+import { ApiProperty } from "@nestjs/swagger";
+
+class CostHistory {
+  @ApiProperty()
+  intervalStart: string;
+  @ApiProperty()
+  totalCost: number;
+}
+
+export class CostHistoryDto {
+  @ApiProperty({
+    type: [CostHistory],
+  })
+  costHistory: CostHistory[];
+}