+++ apps/ui/app/pages/dashboard/cards/AnalyticsCard.tsx
@@ -1,15 +1,16 @@
 import { Card, CardContent, CardHeader, CardTitle } from "~/components/ui/card";
 import { LucideIcon } from "lucide-react";
+import { ReactNode } from "react";
 
 type AnalyticsCardProps = {
   title: string;
   icon: LucideIcon;
-  content: () => JSX.Element;
+  children: ReactNode;
 }
 
-export const AnalyticsCard = ({ title, icon: Icon, content: Content }: AnalyticsCardProps) => {
+export const AnalyticsCard = ({ title, icon: Icon, children }: AnalyticsCardProps) => {
   return (
-    <Card className="flex flex-col p-8 flex-grow flex-1">
+    <Card className="flex flex-col flex-1 flex-grow p-8">
       <CardHeader className={"p-0"}>
         <CardTitle>
           <div className={"flex justify-between font-medium"}>
@@ -19,7 +20,7 @@ export const AnalyticsCard = ({ title, icon: Icon, content: Content }: Analytics
         </CardTitle>
       </CardHeader>
       <CardContent className={"p-0 mt-2 flex-grow"}>
-        <Content />
+        {children}
       </CardContent>
     </Card>
   );