+++ apps/api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # actions
 
+## 1.1.30
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/db@1.1.30
+
 ## 1.1.29
 
 ### Patch Changes