+++ apps/montelo/CHANGELOG.md
@@ -1,5 +1,16 @@
 # montelo
 
+## 1.1.30
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/cli@1.1.30
+  - @montelo/common@1.1.30
+  - @montelo/fizz@1.1.30
+  - @montelo/promptfile@1.1.30
+
 ## 1.1.29
 
 ### Patch Changes